pyproject.toml
[project]
name = "dbsql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-debug-toolbar>=5.2.0",
    "django-environ>=0.12.0",
    "django-resized>=1.0.3",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg>=3.2.9",
    "python-slugify>=8.0.4",
    "redis>=6.2.0",
]


[tool.ruff]
line-length = 79

[tool.ruff.format]
quote-style = "single"
